refactor: 게임 재시작 여부 확인 & 사용자에게 입력을 받는 부분 메서드 추출
사용자가 정답을 맞추었을 때,
게임을 재시작 할 것인지 종료 할 것인지에 대한 여부를 확인할 수 있는 부분을
conRestart 메서드로 추출함
동시에 사용자의 입력 값을 검사하는 메서드도 이동
