+++ src/main/java/baseball/Application.java
@@ -12,26 +12,15 @@ public class Application {
         Game game = new Game();
         do {
             game.start();
-            String userInput; // refactoring 중 임시 변수
+
             if (!game.isAnswer()) {
                 continue;
             }
-            PrintComment.printRestartComment();
-            userInput = Console.readLine();
-
-            validUserInputRestart(userInput);
-            PrintComment.printUserInput(userInput);
-            if (userInput.equals(END)){
+            if (!game.canRestart()){
                 isContinue = false;
                 continue;
             }
             game = new Game();
         } while(isContinue);
-    }
-
-    public static void validUserInputRestart(String num) {
-        if (!num.equals(RESTART) && !num.equals(END)) {
-            throw new IllegalArgumentException(INVALID_INPUT_VALUE);
-        }
     }
 }
